feat: improve type conversions to raise compile-time warnings on truncation
Resolves #762

+++ example/glide_computer_lib/include/glide_computer_lib.h
@@ -102,7 +102,7 @@ struct weather {
 
 struct waypoint {
   std::string name;
-  geographic::position<long double> pos;
+  geographic::position<double> pos;
   geographic::msl_altitude alt;
 };
